refactor: move process record to a separate file
